- Entry point asset manifest model and tests
