+++ tests/Assets/Model/EntryAssetTest.php
@@ -2,7 +2,7 @@
 
 namespace Assets\Model;
 
-use Kanopi\Components\Assets\Model\EntryPoint;
+use Kanopi\Components\Assets\Model\EntryAsset;
 use PHPUnit\Framework\TestCase;
 
 /**
@@ -16,69 +16,67 @@ class EntryAssetTest extends TestCase {
 	 *
 	 * @return array[]
 	 */
-	public function providerArrayType(): array {
+	public function providerScriptPath(): array {
 		return [
-			'Invalid Type - Is a script'  => [
-				'sample-default',
-				[
-					'path' => './path/to/register-only.css',
-					'type' => 'something-invalid',
-				],
-				'script',
+			'Invalid - No Script Path' => [
+				'theme',
+				[],
+				null,
 			],
-			'Missing Type - Auto-detects' => [
-				'sample-default',
+			'Valid - Script Path'      => [
+				'theme',
 				[
-					'path' => './path/to/not-a-script.css',
+					'js' => 'js/theme.1b08d747a2b3b1a1e1ad.js',
 				],
-				'style',
-			],
-			'Valid Type'                  => [
-				'sample-default',
-				[
-					'path' => './path/to/register-only.css',
-					'type' => 'register-only-style',
-				],
-				'register-only-style',
+				'js/theme.1b08d747a2b3b1a1e1ad.js',
 			],
 		];
 	}
 
 	/**
-	 * Test data provider for string versions
+	 * Test data provider for style paths
 	 *
 	 * @return array[]
 	 */
-	public function providerStringType(): array {
+	public function providerStylePath(): array {
 		return [
-			'CSS File'  => [ 'sample-css', './path/to/whatever.css', 'style' ],
-			'JS File'   => [ 'sample-js', './path/to/whatever.js', 'script' ],
-			'JSX File'  => [ 'sample-jsx', './path/to/whatever.jsx', 'script' ],
-			'SASS File' => [ 'sample-sass', './path/to/whatever.sass', 'style' ],
-			'SCSS File' => [ 'sample-scss', './path/to/whatever.scss', 'style' ],
-			'TSX File'  => [ 'sample-tsx', './path/to/whatever.tsc', 'script' ],
+			'Invalid - No Style Path' => [
+				'editor',
+				[
+					'js' => 'js/editor.1b08d747a2b3b1a1e1ad.js',
+				],
+				null,
+			],
+			'Valid - Style Path'      => [
+				'editor',
+				[
+					'css' => 'css/editor.5a1c48eeabe59064833c.css',
+					'js'  => 'js/editor.1b08d747a2b3b1a1e1ad.js',
+				],
+				'css/editor.5a1c48eeabe59064833c.css',
+			],
 		];
 	}
 
 	/**
-	 * @dataProvider providerArrayType
+	 * @dataProvider providerScriptPath
 	 *
-	 * @param string $handle       Entry point handle
-	 * @param array  $entry        Entry configuration array
-	 * @param string $expectedType Expected entry type
+	 * @param string      $entryName     Entry point name
+	 * @param array       $manifestPaths Entry manifest path array
+	 * @param string|null $expected      Expected path
 	 */
-	public function testCheckArrayBuiltType( string $handle, array $entry, string $expectedType ): void {
-		$this->assertEquals( $expectedType, EntryPoint::fromArray( $handle, $entry )->type() );
+	public function testScriptPath( string $entryName, array $manifestPaths, ?string $expected ): void {
+		$this->assertEquals( $expected, EntryAsset::fromArray( $entryName, $manifestPaths )->script() );
 	}
 
 	/**
-	 * @dataProvider providerStringType
+	 * @dataProvider providerStylePath
 	 *
-	 * @param string $handle       Entry point handle
-	 * @param string $path         Source file path
-	 * @param string $expectedType Expected entry type
+	 * @param string      $entryName     Entry point name
+	 * @param array       $manifestPaths Entry manifest path array
+	 * @param string|null $expected      Expected path
 	 */
-	public function testCheckStringBuiltType( string $handle, string $path, string $expectedType ): void {
-		$this->assertEquals( $expectedType, EntryPoint::fromString( $handle, $path )->type() );
+	public function testStylePath( string $entryName, array $manifestPaths, ?string $expected ): void {
+		$this->assertEquals( $expected, EntryAsset::fromArray( $entryName, $manifestPaths )->style() );
 	}
 }
