+++ tests/Assets/Transformers/JavaScriptConfigurationParserTest.php
@@ -63,7 +63,7 @@ class JavaScriptConfigurationParserTest extends TestCase {
 		$readLength  = $readEnd->diff( $start );
 		$isReadLong  = 0 < $readLength->h || 0 < $readLength->i || 0 < $readLength->s || .002 < $readLength->f;
 		$parseLength = $parseEnd->diff( $readEnd );
-		$isParseLong = 0 < $parseLength->h || 0 < $parseLength->i || 0 < $parseLength->s || .001 < $parseLength->f;
+		$isParseLong = 0 < $parseLength->h || 0 < $parseLength->i || 0 < $parseLength->s || .002 < $parseLength->f;
 
 		$this->assertIsArray( $configuration );
 		$this->assertIsArray( $configuration['devServer'] );
@@ -71,6 +71,6 @@ class JavaScriptConfigurationParserTest extends TestCase {
 		$this->assertEquals( $_devServerError, isset( $configuration['devServer']['error'] ), 'devServer invalid' );
 		$this->assertEquals( $_filePatternsError, isset( $configuration['filePatterns']['error'] ), 'filePatterns invalid' );
 		$this->assertFalse( $isReadLong, 'Read longer than 2 ms: ' . $readLength->format( '%h:%i:%s.%f' ) );
-		$this->assertFalse( $isParseLong, 'Parse longer than 1 ms: ' . $parseLength->format( '%h:%i:%s.%f' ) );
+		$this->assertFalse( $isParseLong, 'Parse longer than 2 ms: ' . $parseLength->format( '%h:%i:%s.%f' ) );
 	}
 }
